fix: Fix main on mobile (#14)
* fix: Fix main in mobile

* refactor: Change devDependencies and dependencies
