--- src/pages/main/main.module.ts
@@ -1,11 +1,0 @@
-import axios from 'axios';
-
-import { getMainMarkdownConfig } from '../../lib';
-
-export const getMarkdownDocument = async (): Promise<string> => {
-  const {
-    data: { content: source },
-  } = await axios(getMainMarkdownConfig());
-
-  return source;
-};
